ztev/client/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# client/app.py: 3
# client/config.py: 2
-e git+git://github.com/piyushroshan/flask-oidc-cognito.git@master#egg=flask_oidc_cognito

# client/app.py: 2
# client/config.py: 1,5
Flask == 1.1.2

# client/config.py: 6
Flask_Bootstrap == 3.3.7.1

# client/config.py: 4
Flask_RESTful == 0.3.8

# client/app.py: 5,6
python_jose == 3.2.0

# client/app.py: 10
requests == 2.25.1

# client/config.py: 3
sqlitedict == 1.7.0
